Rename Profile -> PublicProfile for clarity
